reorganize entity schemas
